zendesk_app/requirements.txt
Django==1.7
argparse==1.2.1
django-cors-headers==0.13
djangorestframework==2.4.2
pep8==1.5.7
requests==2.4.1
wsgiref==0.1.2
